Rename basic_algorithm.py to basic_algorithms.py
