+++ opto/trainer/algorithms/__init__.py
@@ -1,1 +1,1 @@
-from opto.trainer.algorithms.basic_algorithm import Minibatch, MinibatchAlgorithm, BasicSearchAlgorithm
+from opto.trainer.algorithms.basic_algorithms import Minibatch, MinibatchAlgorithm, BasicSearchAlgorithm
