+++ opto/trainer/algorithms/basic_algorithms.py
